Radical update of stepwavetest.py
This version has major changes as part of making this the basis for a proper stepper driver.

There is now a class that represents a single motor, and the various settings (pin numbers primarily) are parameters to the constructors.

The motor control code runs in a separate thread with a simple queue on which commands can be sent. An additional thread monitors the return queue and prints responses, allowing the user to type further commands.